+++ Ruby-Proyect/app/views/dashboard/products/edit.html.erb
@@ -31,7 +31,7 @@
 
   <div>
     <%= form.label :imagen %>
-    <%= form.file_field :imagen, value: @product.imagen %>
+    <%= form.file_field :imagen, value: @product.imagen, accept: 'image/*', id: 'imagen-input' %>
   </div>
 
   <div>
@@ -48,3 +48,19 @@
   <%= link_to 'Volver', dashboard_products_path %>
 <% end %>
 
+<script>
+  document.addEventListener("DOMContentLoaded", function () {
+    const input = document.getElementById("imagen-input");
+
+    if (input) {
+      input.addEventListener("change", function () {
+        const file = this.files[0];
+
+        if (file && !file.type.startsWith("image/")) {
+          alert("Por favor selecciona solo archivos de imagen.");
+          this.value = ""; // limpia el campo
+        }
+      });
+    }
+  });
+</script>
