+++ Ruby-Proyect/app/views/dashboard/products/index.html.erb
@@ -1,25 +1,31 @@
+<!-- app/views/dashboard/products/index.html.erb -->
 <h2>Lista de Productos</h2>
 
 <div class="productss">
   <% @products.each do |product| %>
     <div class="product-card">
       <h3><%= product.nombre %></h3>
+      <%= image_tag(product.imagen.attached? ? product.imagen : "LogoLogoText2.svg") %>
       <p>id: <%= product.id %></p>
       <p><strong>Ingredientes:</strong> <%= product.ingredientes.join(", ") %></p>
       <p>Descripción: <%= product.descripcion %></p>
-      <p>Precio: $<%= product.precio %></p>
+      <p>Precio: <%= number_to_currency(product.precio, unit: "COP ", separator: ",", delimiter: ".", precision: 0) %> $</p>
 
-      <div class="estado">
-        <% if product.disponible %>
-          <%= button_to '', toggle_disponibilidad_dashboard_product_path(product), method: :patch, class: 'btn-on' %>
-        <% else %>
-          <%= button_to '', toggle_disponibilidad_dashboard_product_path(product), method: :patch, class: 'btn-off' %>
-        <% end %>
+      <div id="estado-<%= product.id %>">
+        <%= button_to '',
+          toggle_disponibilidad_dashboard_product_path(product),
+          method: :patch,
+          data: { turbo_stream: true }, 
+          class: (product.disponible ? 'btn-on' : 'btn-off'),
+          form: { data: { turbo: true } } 
+        %>
       </div>
+
+
+
 
       <%= link_to "Editar", edit_dashboard_product_path(product), class: "btn-editar" %>
       <%= button_to "Eliminar", dashboard_product_path(product), method: :delete, class: "btn-eliminar" %>
-
     </div>
   <% end %>
 </div>
