+++ Ruby-Proyect/app/views/dashboard/products/toggle_disponibilidad.turbo_stream.erb
@@ -1,0 +1,10 @@
+<%= turbo_stream.replace "estado-#{@product.id}" do %>
+  <div id="estado-<%= @product.id %>">
+    <%= button_to '', toggle_disponibilidad_dashboard_product_path(@product),
+      method: :patch,
+      data: { turbo_stream: true },
+      class: (@product.disponible ? 'btn-on' : 'btn-off'),
+      form: { data: { turbo: true } } 
+    %>
+  </div>
+<% end %>
