+++ Ruby-Proyect/app/views/layouts/application.html.erb
@@ -23,9 +23,8 @@
   </head>
 
   <body>
+
     <%= yield %>
-
-    <%= javascript_include_tag "application", "data-turbo-track": "reload" %>
-
+    
   </body>
 </html>
