+++ Ruby-Proyect/app/views/layouts/dashboard.html.erb
@@ -5,7 +5,7 @@
   <%= csrf_meta_tags %>
   <%= csp_meta_tag %>
   <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
-  <%= javascript_include_tag "application", "data-turbo-track": "reload", defer: true %>
+  <%= javascript_importmap_tags %>
 
 
   <link rel="icon" href="<%= asset_path('LogoLogo.svg') %>" type="image/svg+xml">
@@ -49,5 +49,6 @@
     </main>
 
   </div>
+  <div id="animations"></div>
 </body>
 </html>
