+++ Ruby-Proyect/bin/setup
