+++ Ruby-Proyect/bin/thrust
