+++ Ruby-Proyect/config/importmap.rb
@@ -1,7 +1,5 @@
-# Pin npm packages by running ./bin/importmap
-
 pin "application"
-pin "@hotwired/turbo-rails", to: "turbo.min.js"
-pin "@hotwired/stimulus", to: "stimulus.min.js"
-pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
-pin_all_from "app/javascript/controllers", under: "controllers"
+pin "@hotwired/turbo-rails", to: "turbo.min.js", preload: true
+pin "@hotwired/turbo", to: "turbo.js" # opcional, puede eliminarse si no usas directo esto
+pin "@rails/actioncable/src", to: "@rails--actioncable--src.js" # si lo necesitas
+pin "script", preload: true
